:art: repo sync

+++ app/src/main/java/org/b3log/siyuan/Repo.java
@@ -21,6 +21,7 @@ import org.eclipse.jgit.transport.PushResult;
 import org.eclipse.jgit.transport.SshSessionFactory;
 import org.eclipse.jgit.transport.SshTransport;
 import org.eclipse.jgit.util.FS;
+import org.eclipse.jgit.util.StringUtils;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
@@ -38,6 +39,7 @@ public final class Repo {
     @JavascriptInterface
     public void sync() {
         Toast.makeText(activity, Androidk.language(22), Toast.LENGTH_SHORT).show();
+        String keyFile = null;
         try {
             Androidk.prepareSync();
 
@@ -45,10 +47,7 @@ public final class Repo {
             final String confStr = FileUtils.readFileToString(new File(siyuan + "/conf/conf.json"));
             final JSONObject conf = new JSONObject(confStr);
             final JSONArray boxes = conf.optJSONArray("boxes");
-
-            final String keyFile = Androidk.genTempKeyFile();
-            KEY_FILE = keyFile;
-
+            keyFile = Androidk.genTempKeyFile();
             for (int i = 0; i < boxes.length(); i++) {
                 final JSONObject box = boxes.getJSONObject(i);
                 if (box.optBoolean("isRemote")) {
@@ -58,35 +57,66 @@ public final class Repo {
                 final String localPath = box.optString("path");
                 final Git repo = Git.open(new File(localPath));
 
-                pull(repo);
+                pull(repo, keyFile);
                 Androidk.reloadBox(localPath);
-                push(repo);
-
+                push(repo, keyFile);
+                repo.close();
                 Log.i("", "synced box [" + box.optString("name") + "]");
             }
 
             Androidk.reloadRecentBlocks();
-            FileUtils.deleteQuietly(new File(keyFile));
         } catch (final Exception e) {
             Toast.makeText(activity, e.getMessage(), Toast.LENGTH_SHORT).show();
+        } finally {
+            if (!StringUtils.isEmptyOrNull(keyFile)) {
+                FileUtils.deleteQuietly(new File(keyFile));
+            }
         }
     }
 
-    private static void pull(final Git repo) throws Exception {
+    private static void pull(final Git repo, final String keyFile) throws Exception {
+        final SshSessionFactory jschConfigSessionFactory = new JschConfigSessionFactory() {
+            @Override
+            protected void configure(final OpenSshConfig.Host host, final Session session) {
+                session.setConfig("StrictHostKeyChecking", "no");
+            }
+
+            @Override
+            protected JSch createDefaultJSch(final FS fs) throws JSchException {
+                final JSch defaultJSch = super.createDefaultJSch(fs);
+                defaultJSch.addIdentity(keyFile);
+                return defaultJSch;
+            }
+        };
+
         final PullCommand pullCommand = repo.pull();
         pullCommand.setTransportConfigCallback(transport -> {
-            SshTransport sshTransport = (SshTransport) transport;
-            sshTransport.setSshSessionFactory(JSCH_CONFIG_SESSION_FACTORY);
+            final SshTransport sshTransport = (SshTransport) transport;
+            sshTransport.setSshSessionFactory(jschConfigSessionFactory);
         });
         final PullResult pullResult = pullCommand.call();
         Log.i("", pullResult.toString());
     }
 
-    private static void push(final Git repo) throws Exception {
+    private static void push(final Git repo, final String keyFile) throws Exception {
+        final SshSessionFactory jschConfigSessionFactory = new JschConfigSessionFactory() {
+            @Override
+            protected void configure(final OpenSshConfig.Host host, final Session session) {
+                session.setConfig("StrictHostKeyChecking", "no");
+            }
+
+            @Override
+            protected JSch createDefaultJSch(final FS fs) throws JSchException {
+                final JSch defaultJSch = super.createDefaultJSch(fs);
+                defaultJSch.addIdentity(keyFile);
+                return defaultJSch;
+            }
+        };
+
         final PushCommand pushCommand = repo.push();
         pushCommand.setTransportConfigCallback(transport -> {
-            SshTransport sshTransport = (SshTransport) transport;
-            sshTransport.setSshSessionFactory(JSCH_CONFIG_SESSION_FACTORY);
+            final SshTransport sshTransport = (SshTransport) transport;
+            sshTransport.setSshSessionFactory(jschConfigSessionFactory);
         });
 
         final Iterable<PushResult> pushResults = pushCommand.call();
@@ -94,20 +124,4 @@ public final class Repo {
             Log.i("", result.toString());
         }
     }
-
-    private static String KEY_FILE;
-
-    private static final SshSessionFactory JSCH_CONFIG_SESSION_FACTORY = new JschConfigSessionFactory() {
-        @Override
-        protected void configure(final OpenSshConfig.Host host, final Session session) {
-            session.setConfig("StrictHostKeyChecking", "no");
-        }
-
-        @Override
-        protected JSch createDefaultJSch(final FS fs) throws JSchException {
-            final JSch defaultJSch = super.createDefaultJSch(fs);
-            defaultJSch.addIdentity(KEY_FILE);
-            return defaultJSch;
-        }
-    };
 }
